% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summary-tsco.R
\name{.tsco_print_diagnostics}
\alias{.tsco_print_diagnostics}
\title{Print convergence and separation diagnostics, when there are any.}
\usage{
.tsco_print_diagnostics(diagnostics)
}
\arguments{
\item{diagnostics}{The \code{diagnostics} element of a \code{summary.tsco} object.}
}
\description{
Print convergence and separation diagnostics, when there are any.
}
