% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summary-tsco.R
\name{print.summary.tsco}
\alias{print.summary.tsco}
\title{Print a TSCO summary}
\usage{
\method{print}{summary.tsco}(
  x,
  digits = max(3L, getOption("digits") - 3L),
  signif.stars = getOption("show.signif.stars"),
  ...
)
}
\arguments{
\item{x}{An object of class \code{"summary.tsco"}.}

\item{digits}{Number of digits to print.}

\item{signif.stars}{Logical; whether to print significance stars.}

\item{...}{Additional arguments passed to print methods.}
}
\description{
Print a TSCO summary
}
