% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/print-tsco.R
\name{print.tsco}
\alias{print.tsco}
\title{Print a TSCO model}
\usage{
\method{print}{tsco}(x, ...)
}
\arguments{
\item{x}{An object of class \code{"tsco"}.}

\item{...}{Ignored.}
}
\description{
Print a TSCO model
}
