% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/logLik-tsco.R
\name{logLik.tsco}
\alias{logLik.tsco}
\title{Extract log-likelihood from a TSCO model}
\usage{
\method{logLik}{tsco}(object, ...)
}
\arguments{
\item{object}{An object of class \code{"tsco"}.}

\item{...}{Ignored.}
}
\value{
An object of class \code{"logLik"}.
}
\description{
Extracts the constant-free TSCO log-likelihood, computed on the
individual-level scale. For grouped/count responses, this excludes
multinomial combinatorial constants so that grouped and ungrouped
representations of the same data have comparable log-likelihoods.
}
