% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summary-tsco.R
\name{summary.tsco}
\alias{summary.tsco}
\title{Summarize a TSCO model}
\usage{
\method{summary}{tsco}(object, ...)
}
\arguments{
\item{object}{An object of class \code{"tsco"}.}

\item{...}{Ignored.}
}
\value{
An object of class \code{"summary.tsco"}.
}
\description{
Produces a concise statistical summary of both fitted TSCO stages.
}
