% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tsco-package.R
\docType{package}
\name{tsco-package}
\alias{tsco-package}
\title{tsco: Two-Stage Conditional Odds Models}
\description{
Implements two-stage conditional odds (TSCO) models for ordinal outcomes using VGAM. Supports individual-level ordinal responses and grouped count responses.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/psboonstra/tsco}
  \item Report bugs at \url{https://github.com/psboonstra/tsco/issues}
}

}
\author{
\strong{Maintainer}: Philip S. Boonstra \email{phil@umich.edu}

Authors:
\itemize{
  \item Philip S. Boonstra \email{phil@umich.edu}
}

}
\keyword{internal}
