% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vcov-tsco.R
\name{vcov.tsco}
\alias{vcov.tsco}
\title{Variance-covariance matrix for a TSCO model}
\usage{
\method{vcov}{tsco}(object, stage = c("both", "stage1", "stage2"), ...)
}
\arguments{
\item{object}{An object of class \code{"tsco"}.}

\item{stage}{Which stage to extract: \code{"both"}, \code{"stage1"}, or \code{"stage2"}.}

\item{...}{Ignored.}
}
\value{
A variance-covariance matrix. If \code{stage = "both"}, the returned
matrix is block diagonal with stage-specific coefficient names prefixed by
\code{"stage1:"} and \code{"stage2:"}.
}
\description{
Variance-covariance matrix for a TSCO model
}
